+++ backend/sql/__mocks__/database.js
@@ -33,14 +33,7 @@ const mockDatabase = {
     updatePointCoordinates: jest.fn().mockResolvedValue(true),
     updatePointNorthDirection: jest.fn().mockResolvedValue(true),
     updatePointImage: jest.fn().mockResolvedValue(true),
-    getConnectionsByGameMapId: jest.fn(),
-    deleteConnectionById: jest.fn().mockResolvedValue(true),
-    arePointsInSameGameMap: jest.fn().mockResolvedValue(true),
-    doesConnectionAlreadyExist: jest.fn().mockResolvedValue(false),
     insertConnection: jest.fn(),
-    arePointsInSameMap: jest.fn().mockResolvedValue(true),
-    isConnectionCrossMap: jest.fn().mockResolvedValue(true),
-    updateConnectionDirections: jest.fn().mockResolvedValue(true),
 
     getGameMapComments: jest.fn(),
     getGameMapCommentCount: jest.fn(),
