+++ backend/sql/database.js
@@ -9,6 +9,7 @@ const bcrypt = require('bcrypt');
 //     return rows;
 // }
 
+// TODO: eltavolitani miutan ki lett veve minden fuggveny ami hasznalta
 function isIdUpdateSuccessful(result) {
     const match = result?.info?.match(/Rows matched:\s*(\d+)/);
     const rowsMatched = match ? parseInt(match[1]) : 0;
@@ -479,17 +480,6 @@ async function insertPoint(connection, mapId, u, v, northDirection, imageId) {
     return result.insertId;
 }
 
-async function insertConnection(connection, startPointId, endPointId, gameMapId, startToEndDirection, endToStartDirection) {
-    const query = `
-        INSERT INTO point_connections (start_point_id, end_point_id, game_maps_id, direction_start_to_end, direction_end_to_start)
-        VALUES (?, ?, ?, ?, ?)
-    `;
-    const smallerId = Math.min(startPointId, endPointId);
-    const largerId = Math.max(startPointId, endPointId);
-    const [result] = await connection.execute(query, [smallerId, largerId, gameMapId, startToEndDirection, endToStartDirection]);
-    return result.insertId;
-}
-
 async function updateImagePath(connection, imageId, filepath) {
     const query = `
         UPDATE images
@@ -530,25 +520,6 @@ async function getPointsOnMap(mapId) {
         WHERE map.map_id = ?
     `;
     const [rows] = await pool.execute(query, [mapId]);
-    return rows;
-}
-
-async function getConnectionsByGameMapId(gameMapId) {
-    const query = `
-        SELECT
-            point_connections.connection_id,
-            point_connections.start_point_id,
-            point_connections.end_point_id,
-            start_point.map_id AS start_map_id,
-            end_point.map_id AS end_map_id,
-            point_connections.direction_start_to_end,
-            point_connections.direction_end_to_start
-        FROM point_connections
-            INNER JOIN points AS start_point ON (start_point.point_id = point_connections.start_point_id)
-            INNER JOIN points AS end_point ON (end_point.point_id = point_connections.end_point_id)
-        WHERE point_connections.game_maps_id = ?
-    `;
-    const [rows] = await pool.execute(query, [gameMapId]);
     return rows;
 }
 
@@ -738,43 +709,6 @@ async function deleteImageById(connection, imageId) {
     return result.affectedRows == 1;
 }
 
-async function arePointsInSameGameMap(connection, pointId1, pointId2, gameMapId) {
-    const query = `
-        SELECT COUNT(DISTINCT points.point_id) AS count
-        FROM points
-            INNER JOIN map ON (points.map_id = map.map_id)
-        WHERE points.point_id IN (?, ?) 
-          AND map.game_maps_id = ?;
-    `;
-
-    const [rows] = await connection.execute(query, [pointId1, pointId2, gameMapId]);
-
-    return rows[0].count == 2;
-}
-
-async function arePointsInSameMap(connection, pointId1, pointId2) {
-    const query = `
-        SELECT COUNT(DISTINCT map_id) AS map_count
-        FROM points
-        WHERE point_id IN (?, ?);
-    `;
-
-    const [rows] = await connection.execute(query, [pointId1, pointId2]);
-
-    return rows[0].map_count == 1;
-}
-
-async function doesConnectionAlreadyExist(connection, pointId1, pointId2) {
-    const query = `
-        SELECT COUNT(*) as count 
-        FROM point_connections
-        WHERE
-            (point_connections.start_point_id, point_connections.end_point_id) IN ((?, ?), (?, ?))
-    `;
-    const [rows] = await connection.execute(query, [pointId1, pointId2, pointId2, pointId1]);
-    return rows[0].count == 1;
-}
-
 async function deletePointById(connection, pointId) {
     const query = `
         DELETE FROM points
@@ -791,40 +725,6 @@ async function deleteMapById(connection, mapId) {
     `;
     const [result] = await connection.execute(query, [mapId]);
     return result.affectedRows == 1;
-}
-
-async function deleteConnectionById(connection, connectionId) {
-    const query = `
-        DELETE FROM point_connections
-        WHERE point_connections.connection_id = ?
-    `;
-    const [result] = await connection.execute(query, [connectionId]);
-    return result.affectedRows == 1;
-}
-
-async function updateConnectionDirections(connection, connectionId, dirStartToEnd, dirEndToStart) {
-    const query = `
-        UPDATE point_connections
-        SET direction_start_to_end = COALESCE(?, direction_start_to_end),
-            direction_end_to_start = COALESCE(?, direction_end_to_start)
-        WHERE connection_id = ?
-    `;
-    const [result] = await connection.execute(query, [dirStartToEnd, dirEndToStart, connectionId]);
-    return isIdUpdateSuccessful(result);
-}
-
-async function isConnectionCrossMap(connection, connectionId) {
-    const query = `
-        SELECT 
-            start_point.map_id AS start_map_id,
-            end_point.map_id AS end_map_id
-        FROM point_connections
-            INNER JOIN points start_point ON (start_point.point_id = point_connections.start_point_id)
-            INNER JOIN points end_point ON (end_point.point_id = point_connections.end_point_id)
-        WHERE point_connections.connection_id = ?
-    `;
-    const [rows] = await connection.execute(query, [connectionId]);
-    return rows[0].start_map_id != rows[0].end_map_id;
 }
 
 async function getGameMapDetails(gameMapID) {
@@ -981,7 +881,7 @@ async function updateUserCommentOnGameMap(connection, gameMapId, userId, comment
         WHERE game_maps_comments.game_maps_id = ? AND game_maps_comments.user_id = ?
     `;
     const [result] = await connection.execute(query, [commentText, rating, gameMapId, userId]);
-    
+
     return isIdUpdateSuccessful(result);
 }
 
@@ -1023,12 +923,10 @@ module.exports = {
     insertImage,
     insertMap,
     insertPoint,
-    insertConnection,
     updateImagePath,
     getMapImage,
     getPointImage,
     getPointsOnMap,
-    getConnectionsByGameMapId,
     getConnectionsByPointId,
     getMapsByGameMapId,
     getGameMapIdByMapId,
@@ -1047,11 +945,8 @@ module.exports = {
     newUserFromAdmin,
     getUserByUsername,
     getUserByEmail,
-    arePointsInSameGameMap,
-    doesConnectionAlreadyExist,
     deletePointById,
     deleteMapById,
-    deleteConnectionById,
     getUsers,
     getUser,
     getUserNameProfile,
@@ -1067,9 +962,6 @@ module.exports = {
     getMapImageIdByMapId,
     getMapInfo,
     getAllImageIdsForMap,
-    arePointsInSameMap,
-    updateConnectionDirections,
-    isConnectionCrossMap,
     getGameMapDetails,
     getTopScoresForGameMap,
     doesGameMapExist,
@@ -1084,5 +976,6 @@ module.exports = {
     updateUserCommentOnGameMap,
     deleteUserCommentOnGameMap,
     getAllImageIdsForGameMap,
-    deleteGameMapById
+    deleteGameMapById,
+    isIdUpdateSuccessful
 };
