+++ backend/test/jest.setup.js
@@ -3,3 +3,4 @@ jest.mock("#utils/image-processor.js");
 jest.mock("#utils/file-utils.js");
 jest.mock("#root/auth.js");
 jest.mock("#gamemaps/shared/middlewares/gamemaps.middleware.js");
+jest.mock("#mapcreator/connections/connections.queries.js");
